Configuration for all environments to use DateTime
In one of the previous PR's I had the misconception that I could delcare
the time_api and logger config just at the top layer and in test. This
is not the case however after trying to implement this into a library I
received nil for the :time_api unless I implemented that in the project
itself, this should solve this problem.