Use non ! Jason.encode to save parent app
When using the ! for both Jason and Poison .encode! the error would be
raised causing the Logger to fail, the result of this was that the
parent app using Stump could crash. This change should solve that
problem ensuring we dont forcibly raise the exception but instead log
the error message we receive when trying to encode it.